requirements.txt
# Core dependencies (v1 and v2)
requests>=2.28.0
openai>=1.0.0
click>=8.0.0

# V2 RAG pipeline dependencies
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# Optional: vLLM for local LLM serving
# Uncomment if you want to use vLLM:
# vllm>=0.2.0
